+++ srcs/wolf_fog.c
@@ -6,7 +6,7 @@
 /*   By: tmaluh <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/16 19:10:46 by tmaluh            #+#    #+#             */
-/*   Updated: 2019/02/18 15:17:29 by tmaluh           ###   ########.fr       */
+/*   Updated: 2019/02/19 13:37:22 by tmaluh           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,8 +22,8 @@ Uint32	wolf_fog_change(t_clrs *c)
 	return (colors[*c]);
 }
 
-Uint32	wolf_fog(double dist, Uint32 src_color,
-				Uint32 fog_color, double max_fog_dist)
+Uint32	wolf_fog(float dist, Uint32 src_color,
+				Uint32 fog_color, float max_fog_dist)
 {
 	float		fog_amount;
 	Uint32		dest_color;
